fix it can run without google enviroment variable
