+++ config/index.js
@@ -11,12 +11,13 @@ if (process.env.NODE_ENV !== 'production') {
 const db = require('./db')
 const web = require('./web')
 const secret = require('./secret')
-
+const google = require('./google')
 
 const config = {
   db,
   web,
-  secret
+  secret,
+  google
 }
 
 class ConfigManager {
