+++ routes/user.js
@@ -5,39 +5,20 @@ const router = express.Router()
 const { dataSource } = require('../db/data-source')
 const handleErrorAsync = require('../utils/handleErrorAsync')
 const userController = require('../controllers/user')
-const { generateJWT } = require('../utils/jwtUtils');
-const { createOrLoginGoogleAccount } = require('../services/userService')
-
-const { isValidString } = require('../utils/validUtils');
-const { isValidUrl } = require('../utils/validUtils');
-const { isUndefined } = require('../utils/validUtils');
 const isAuth = require('../middlewares/auth')({
   secret: config.get('secret').jwtSecret,
   userRepository: dataSource.getRepository('User'),
   logger
 })
 
-const passport = require('passport');
-const GoogleStrategy = require( 'passport-google-oauth20' ).Strategy;
-passport.use(new GoogleStrategy({
-    clientID: process.env.GOOGLE_AUTH_CLIENTID,
-    clientSecret: process.env.GOOGLE_AUTH_CLIENT_SECRET,
-    callbackURL: `${ process.env.GOOGLE_AUTH_CALLBACKURL }/api/v1/users/google/callback`,
-    passReqToCallback: true
-  },
-  createOrLoginGoogleAccount
-));
+
 
 // 註冊 & 登入
 router.post('/signup', handleErrorAsync(userController.postSignup));
 router.post('/signin', handleErrorAsync(userController.postLogin));
 
 // GET 所有使用者
-// router.get('/', userController.getAllUsers);
-
-router.get('/', (req, res) => {
-  res.render('index', { title: 'Express', Host: `${ process.env.GOOGLE_AUTH_CALLBACKURL }/api/v1/users/google/signin-or-signup` } );
-});
+router.get('/', userController.getAllUsers);
 
 //  GET 取得使用者資料
 router.get('/profile', isAuth, handleErrorAsync(userController.getProfile));
@@ -49,19 +30,5 @@ router.put('/password', isAuth, handleErrorAsync(userController.putPassword));
 router.post('/auth', isAuth, handleErrorAsync(userController.postAuth));
 // router.get('/auth/refresh', isAuth, handleErrorAsync(userController.getRefresh));
 // router.get('/auth/logout', isAuth, handleErrorAsync(userController.getLogout));
-
-router.get('/google/signin-or-signup', passport.authenticate('google', {
-  scope: [ 'email', 'profile'],
-  state: 'login'
-}));
-
-router.get('/google/bind', isAuth , passport.authenticate('google', {
-  scope: ['email', 'profile'],
-  state: 'bind'
-}));
-
-router.get('/google/callback',
-  passport.authenticate('google', { session: false }),
-  handleErrorAsync(userController.googleCallback))
 
 module.exports = router
